floyd_requirements.txt
Pillow
scipy
numpy
flask
scikit-image
matplotlib
